requirements.txt
flask==3.0.3
cohere==5.9.2
python-dotenv==1.0.1
gunicorn==20.1.0